Tiberium Inhibitor
Addition of Tiberium Inhibitor code.

New XML:
New CompProperties_SonicInhibitor. This can be attached to *any*
building to give it a tiberium destructive effect. This will cause it to
both destroy tiberium crystals, and tiberium plants in the given radius.
Specifically, any plant containing "Tiberium" in the name will be
destroyed. Does not detonate blue tiberium.